Update videolabjqt.ijs to for the new Qt in J9.4
The new jQt will not allow videos to be run from online urls. The media option will run files which are media files but not URL's so the solution is to launch a browser window to run the video outside of jQt the way that the J console does it.